fix(header): Fix forward declaration + norme
