+++ includes/expand.h
@@ -1,7 +1,12 @@
 #ifndef EXPAND_H
 # define EXPAND_H
 
-# include "minishell.h"
+//  *** Forward declarations ***
+struct						s_env;
+typedef struct s_env		t_env;
+
+struct						s_status;
+typedef struct s_status		t_status;
 
 typedef struct s_expand
 {
